Review UI: move toolbar stale/pending notices inline, delete AttentionBanner
The AttentionBanner full-width bar is replaced by inline notices in the
ReviewHeader actions slot (left of the counter picker), matching the
pattern already used in SummaryScreen.

Shared bits extracted to avoid duplication:
- HeaderNoticeText styled export added to ReviewHeader.tsx
- StaleNotice.tsx new component holding STALE_REFRESH_PROMPT and the
  copy-prompt popover content, used by both ReviewToolbarHeader and
  SummaryScreen

AttentionBanner.tsx and AttentionBanner.stories.tsx deleted; coverage
migrated to ReviewPage.stories.tsx with a new ToolbarStale story.
PendingUpdateFromStoryNavigatesToSummary updated: button name changed
from 'Update' (old bar) to 'Refresh review'.
