+++ code/core/src/manager/globals/exports.ts
@@ -239,6 +239,7 @@ export default {
     'ProceedIcon',
     'ProfileIcon',
     'PullRequestIcon',
+    'QRIcon',
     'QuestionIcon',
     'RSSIcon',
     'RedirectIcon',
@@ -261,14 +262,16 @@ export default {
     'SidebarToggleIcon',
     'SortDownIcon',
     'SortUpIcon',
+    'SparkleIcon',
     'SpeakerIcon',
     'StackedIcon',
     'StarHollowIcon',
     'StarIcon',
+    'StatusAddIcon',
     'StatusFailIcon',
     'StatusIcon',
-    'StatusNewIcon',
     'StatusPassIcon',
+    'StatusRemoveIcon',
     'StatusWarnIcon',
     'StickerIcon',
     'StopAltHollowIcon',
